refactor(tests): split service_tests.rs — move flag and machine-rename tests to service_flag_tests.rs
service_tests.rs was 3068 lines, exceeding the new 3000-line linecheck gate
added in the previous commit. Flags (svc_create_flag/list_flags/resolve_flag)
and machine-rename (svc_rename_machine) tests now live in service_flag_tests.rs,
reducing service_tests.rs to 2677 lines while keeping all 760 tests green.
